// models/sales_perfect_binding.model.ts
export type SalesPerfectBindingState = {
  data: {
    sals_perf_id: string;
    sals_perf_date: string;
    sals_perf_company_name: string;
    sals_perf_tel: string;
    sals_perf_email: string;
    sals_perf_doc_type: string;
    sals_perf_printing_type: string;
    sals_perf_amount: string;
    sals_perf_quotation_request: string;
    sals_perf_finished_size: string;
    sals_perf_cover: string;
    sals_perf_text: string;
    sals_perf_cover_paper: string;
    sals_perf_text_paper: string;
    sals_perf_printing: string;
    sals_perf_cover_coating: string;
    sals_perf_text_coating: string;
    sals_perf_printing_volume: string;
  };
};